+++ lib/node/node.h
@@ -11,7 +11,7 @@ class Node {
   uint8_t rampMode;
   uint8_t loopMode;
   uint16_t getCurrentBrightness();
-  void setTarget(uint16_t value, uint16_t duration);
+  void setTarget(uint16_t value, uint32_t duration);
   Node();
 };
